+++ src/main/kotlin/com/github/firusv/smarti18n/dialog/AddDialog.kt
@@ -36,7 +36,6 @@ class AddDialog(
 
     init {
         title = MessagesBundle.message("action.add")
-        translateKeyField.text = ""
     }
 
     /**