+++ CHANGELOG.md
@@ -4,21 +4,27 @@
 
 ## [Unreleased]
 
+## [1.1.2] - 17.12.2024
+
+### Added
+
+- Added context action on translate
+
 ## [1.1.1] - 17.12.2024
 
 ### Added
 
-Added check translation exist
+- Added check translation exist
 
 ## [1.1.0] - 17.12.2024
 
 ### Added
 
-Added go to translate reference by Ctrl+Click 
+- Added go to translate reference by Ctrl+Click 
 
 ### Changes
 
-Added support platform version 243
+- Added support platform version 243
 
 ## [1.0.5] - 12.12.2024